Улучшен компилятор ресурсов

+++ AUI.Core/src/AUI/i18n/i18n.h
@@ -1,6 +1,6 @@
 #pragma once
 
-inline AString operator"" _i18n(const char* input, unsigned long s)
+inline AString operator"" _i18n(const char* input, size_t s)
 {
 	return input;
 }
